README.rd
This is a scatterplot visualising the relationship between fertility and life expectancy based on country data. The size of the circles indicates the total number of births that year.

Tooltips are used to show additional information for each country.

Possible future improvements to this scatterplot include a slider input allowing to change years and adding a color scale to circles to visualise a different aspect of the data (eg. overall population).